refactor: call the right path of schema
