+++ app/routers/endpoints/auth_route.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import select
 from app.models import get_session, User
-from app.schemas import SignUpRequest, LoginRequest, Token
+from app.schemas.user_schema import SignUpRequest, LoginRequest, Token
 from app.core.security import hash_password, verify_password, create_access_token
 
 router = APIRouter(prefix="/auth", tags=["auth"])
