+++ app/routers/endpoints/user_route.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import select
 from app.models import get_session, User
-from app.schemas import MeRead, MeUpdate
+from app.schemas.user_schema import MeRead, MeUpdate
 from app.core.deps import get_current_user
 from app.core.security import verify_password, hash_password
